requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python==1.2.3
